fixed more tests
